hardcode forwarding of XF86_* keys
** !! Absolutely Hazardous !! **

This hardcodes forwarding of specific keys, which suit my need to
control music, while the screen is still locked.

My C skills are exceptionally bad and therefore this commit too.

Actually, the hidden goal of this commit is to annoy the i3lock
maintainers and give them eye cancer while viewing this commit and
therefore they will implement this in i3lock natively.